added policy for task destroy method

+++ app/Http/Controllers/TaskController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Task;
 use Illuminate\Http\Request;
 use App\Repositories\TaskRepository;
 
@@ -33,8 +34,9 @@ class TaskController extends Controller
 		]);
 	}
 
-	public function destroy(Request $request, \App\Task $task)
+	public function destroy(Request $request, Task $task)
 	{
+		$this->authorize('destroy', $task);
 		$task->delete();
 
 		return redirect('/tasks');
